separated kick from commandsChannel
