+++ docs/toDO.txt
@@ -14,6 +14,10 @@ TODO SERVER/ CLIENT/ CHANNEL /ELSE:
 
 
 17.07
+
+BRANCH: "refactor/separate-kick-commands" 
+	separated kick from commandsChannel
+
 BRANCH: "refactor/separate-invite-commands"l 
 	separated invite from commandsChannel
